[IMPYLA-27] Factor out thrift api into internal module
Because of py2 and py3 support, accessing the thrift objects is a bit uglier now.  This commit factors it out into a separate module.

Fixes #75.
